+++ src/CssOklchColor/CssOklchColor.ts
@@ -36,12 +36,15 @@ import { DEFAULT_DATA_PATH, THROW_THE_ERROR, type DataGuaranteeOptions, type Fun
 import type { Oklch } from "culori";
 import type { CssCmykColor } from "../CssCmykColor/CssCmykColor";
 import type { CssCmykColorData } from "../CssCmykColor/CssCmykColorData.type";
+import { makeCssCmykColorFromCssColor } from "../CssCmykColor/makeCssCmykColorFromCssColor";
 import { CssColor } from "../CssColor/CssColor";
 import type { CssHexColorDefinition } from "../CssHexColor/CssHexColorDefinition.type";
 import { CssHslColor } from "../CssHslColor/CssHslColor";
 import type { CssHslColorData } from "../CssHslColor/CssHslColorData.type";
+import { makeCssHslColorFromCssColor } from "../CssHslColor/makeCssHslColorFromCssColor";
 import { CssHwbColor } from "../CssHwbColor/CssHwbColor";
 import type { CssHwbColorData } from "../CssHwbColor/CssHwbColorData.type";
+import { makeCssHwbColorFromCssColor } from "../CssHwbColor/makeCssHwbColorFromCssColor";
 import { CssRgbColor } from "../CssRgbColor/CssRgbColor";
 import type { CssRgbColorData } from "../CssRgbColor/CssRgbColorData.type";
 import { makeCssRgbColorFromCssColor } from "../CssRgbColor/makeCssRgbColorFromCssColor";
@@ -65,7 +68,7 @@ export class CssOklchColor extends CssColor<CssOklchColorData, Oklch>
         ...fnOpts: FunctionalOption<CssCmykColorData, DataGuaranteeOptions>[]
     ): CssCmykColor
     {
-        return this.rgb().cmyk({path, onError}, ...fnOpts);
+        return makeCssCmykColorFromCssColor(this, { path, onError }, ...fnOpts);
     }
 
     public hsl(
@@ -76,9 +79,7 @@ export class CssOklchColor extends CssColor<CssOklchColorData, Oklch>
         ...fnOpts: FunctionalOption<CssHslColorData, DataGuaranteeOptions>[]
     ): CssHslColor
     {
-        // because we are converting between color spaces,
-        // best we go via rgb() first
-        return this.rgb().hsl({ path, onError}, ...fnOpts);
+        return makeCssHslColorFromCssColor(this, { path, onError }, ...fnOpts);
     }
 
     public hwb(
@@ -89,9 +90,7 @@ export class CssOklchColor extends CssColor<CssOklchColorData, Oklch>
         ...fnOpts: FunctionalOption<CssHwbColorData, DataGuaranteeOptions>[]
     ): CssHwbColor
     {
-        // because we are converting between color spaces,
-        // best we go via rgb() first
-        return this.rgb().hwb({ path, onError}, ...fnOpts);
+        return makeCssHwbColorFromCssColor(this, { path, onError }, ...fnOpts);
     }
 
     public oklch(